refactor: split galois8 by arch and runtime role
