+++ src/galois_8/legacy/mod.rs
@@ -1,0 +1,7 @@
+#[cfg(all(
+    feature = "simd-accel",
+    any(target_arch = "x86_64", target_arch = "aarch64"),
+    not(target_env = "msvc"),
+    not(any(target_os = "android", target_os = "ios"))
+))]
+pub(crate) mod simd_c;
