+++ src/galois_8/x86/avx512.rs
@@ -1,0 +1,1 @@
+// Placeholder module for future x86 AVX-512 implementation.
