+++ src/galois_8/x86/gfni.rs
@@ -1,0 +1,1 @@
+// Placeholder module for future x86 GFNI implementation.
