+++ src/galois_8/x86/ssse3.rs
@@ -1,0 +1,1 @@
+// Placeholder module for future x86 SSSE3 implementation.
